move download btn to env-keys component
